Refactor repo suggestions out to a custom hook
